+++ gcs/src/test/java/com/google/cloud/hadoop/fs/gcs/GoogleHadoopFSInputStreamIntegrationTest.java
@@ -17,6 +17,7 @@ package com.google.cloud.hadoop.fs.gcs;
 import static com.google.common.truth.Truth.assertThat;
 import static org.junit.Assert.assertThrows;
 
+import com.google.cloud.hadoop.gcsio.AssertingLogHandler;
 import com.google.cloud.hadoop.gcsio.GoogleCloudStorageFileSystemIntegrationHelper;
 import com.google.cloud.hadoop.gcsio.GoogleCloudStorageReadOptions;
 import java.io.EOFException;
@@ -25,8 +26,13 @@ import java.net.URI;
 import java.nio.channels.ClosedChannelException;
 import java.nio.charset.StandardCharsets;
 import java.util.Arrays;
+import java.util.Map;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 import org.apache.hadoop.fs.FileSystem;
+import org.junit.After;
 import org.junit.AfterClass;
+import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -36,6 +42,8 @@ import org.junit.runners.JUnit4;
 public class GoogleHadoopFSInputStreamIntegrationTest {
 
   private static GoogleCloudStorageFileSystemIntegrationHelper gcsFsIHelper;
+  private AssertingLogHandler assertingHandler = new AssertingLogHandler();
+  private Logger grpcTracingLogger;
 
   @BeforeClass
   public static void beforeClass() throws Exception {
@@ -46,6 +54,19 @@ public class GoogleHadoopFSInputStreamIntegrationTest {
   @AfterClass
   public static void afterClass() {
     gcsFsIHelper.afterAllTests();
+  }
+
+  @Before
+  public void setup() {
+    grpcTracingLogger = Logger.getLogger(GoogleHadoopFSInputStream.class.getName());
+    grpcTracingLogger.setUseParentHandlers(false);
+    grpcTracingLogger.addHandler(assertingHandler);
+    grpcTracingLogger.setLevel(Level.INFO);
+  }
+
+  @After
+  public void cleanUp() {
+    grpcTracingLogger.removeHandler(assertingHandler);
   }
 
   @Test
@@ -63,6 +84,7 @@ public class GoogleHadoopFSInputStreamIntegrationTest {
 
     Throwable exception = assertThrows(EOFException.class, () -> in.seek(testContent.length()));
     assertThat(exception).hasMessageThat().contains("Invalid seek offset");
+    assertingHandler.assertLogCount(0);
   }
 
   @Test
@@ -79,14 +101,57 @@ public class GoogleHadoopFSInputStreamIntegrationTest {
     byte[] value = new byte[2];
     byte[] expected = Arrays.copyOf(testContent.getBytes(StandardCharsets.UTF_8), 2);
 
+    // enabled trace logging
     GoogleCloudStorageReadOptions options =
-        ghfs.getGcsFs().getOptions().getCloudStorageOptions().getReadChannelOptions();
+        GoogleCloudStorageReadOptions.builder().setTraceLogEnabled(true).build();
     FileSystem.Statistics statistics = new FileSystem.Statistics(ghfs.getScheme());
     try (GoogleHadoopFSInputStream in =
         new GoogleHadoopFSInputStream(ghfs, path, options, statistics)) {
       assertThat(in.read(value, 0, 1)).isEqualTo(1);
       assertThat(statistics.getReadOps()).isEqualTo(1);
+
+      assertingHandler.assertLogCount(1);
+      Map<String, Object> logRecord =
+          assertingHandler.getLogRecord(
+              GoogleHadoopFSInputStream.METHOD, GoogleHadoopFSInputStream.READ_METHOD);
+      assertThat(logRecord).isNotNull();
+      assertThat(logRecord.get(GoogleHadoopFSInputStream.BYTES_READ)).isEqualTo(1);
+      assertThat(logRecord.get(GoogleHadoopFSInputStream.POSITION)).isEqualTo(0);
+      assertThat(logRecord.get(GoogleHadoopFSInputStream.DURATION_NS)).isNotNull();
+      assertThat(logRecord.get(GoogleHadoopFSInputStream.GCS_PATH)).isNotNull();
+
+      assertingHandler.flush();
+
       assertThat(in.read(1, value, 1, 1)).isEqualTo(1);
+      // Total 4 calls for positional read
+      // 1. positional Read call
+      // 2. Seek to desired position
+      // 3. actual read from buffer after seek
+      // 4. Seek to old position
+      assertingHandler.assertLogCount(4);
+      logRecord =
+          assertingHandler.getLogRecord(
+              GoogleHadoopFSInputStream.METHOD, GoogleHadoopFSInputStream.POSITIONAL_READ_METHOD);
+      assertThat(logRecord).isNotNull();
+      assertThat(logRecord.get(GoogleHadoopFSInputStream.BYTES_READ)).isEqualTo(1);
+      assertThat(logRecord.get(GoogleHadoopFSInputStream.POSITION)).isEqualTo(1);
+      assertThat(logRecord.get(GoogleHadoopFSInputStream.DURATION_NS)).isNotNull();
+      assertThat(logRecord.get(GoogleHadoopFSInputStream.GCS_PATH)).isNotNull();
+
+      logRecord =
+          assertingHandler.getLogRecord(
+              GoogleHadoopFSInputStream.METHOD, GoogleHadoopFSInputStream.SEEK_METHOD);
+      assertThat(logRecord).isNotNull();
+      assertThat(logRecord.get(GoogleHadoopFSInputStream.POSITION)).isEqualTo(1);
+      assertThat(logRecord.get(GoogleHadoopFSInputStream.DURATION_NS)).isNotNull();
+
+      logRecord =
+          assertingHandler.getLogRecord(
+              GoogleHadoopFSInputStream.METHOD, GoogleHadoopFSInputStream.READ_METHOD);
+      assertThat(logRecord).isNotNull();
+      assertThat(logRecord.get(GoogleHadoopFSInputStream.BYTES_READ)).isEqualTo(1);
+      assertThat(logRecord.get(GoogleHadoopFSInputStream.DURATION_NS)).isNotNull();
+
       assertThat(statistics.getReadOps()).isEqualTo(2);
     }