+++ gcsio/src/test/java/com/google/cloud/hadoop/gcsio/GoogleCloudStorageNewIntegrationTest.java
@@ -110,8 +110,7 @@ public class GoogleCloudStorageNewIntegrationTest {
     list_nonExistentBucket_nullPrefix(
         (gcs, bucket, prefix) -> getObjectNames(gcs.listObjectInfo(bucket, prefix)),
         (bucket, prefix) ->
-            listRequestWithTrailingDelimiter(
-                bucket, prefix, /* maxResults= */ 1024, /* pageToken= */ null));
+            listRequestWithTrailingDelimiter(bucket, prefix, /* pageToken= */ null));
   }
 
   @Test
@@ -121,8 +120,7 @@ public class GoogleCloudStorageNewIntegrationTest {
             getObjectNames(
                 gcs.listObjectInfoPage(bucket, prefix, /* pageToken= */ null).getItems()),
         (bucket, prefix) ->
-            listRequestWithTrailingDelimiter(
-                bucket, prefix, /* maxResults= */ 1024, /* pageToken= */ null));
+            listRequestWithTrailingDelimiter(bucket, prefix, /* pageToken= */ null));
   }
 
   private void list_nonExistentBucket_nullPrefix(
@@ -147,8 +145,7 @@ public class GoogleCloudStorageNewIntegrationTest {
     list_emptyBucket_nullPrefix(
         (gcs, bucket, prefix) -> getObjectNames(gcs.listObjectInfo(bucket, prefix)),
         (bucket, prefix) ->
-            listRequestWithTrailingDelimiter(
-                bucket, prefix, /* maxResults= */ 1024, /* pageToken= */ null));
+            listRequestWithTrailingDelimiter(bucket, prefix, /* pageToken= */ null));
   }
 
   @Test
@@ -158,8 +155,7 @@ public class GoogleCloudStorageNewIntegrationTest {
             getObjectNames(
                 gcs.listObjectInfoPage(bucket, prefix, /* pageToken= */ null).getItems()),
         (bucket, prefix) ->
-            listRequestWithTrailingDelimiter(
-                bucket, prefix, /* maxResults= */ 1024, /* pageToken= */ null));
+            listRequestWithTrailingDelimiter(bucket, prefix, /* pageToken= */ null));
   }
 
   private void list_emptyBucket_nullPrefix(
@@ -184,8 +180,7 @@ public class GoogleCloudStorageNewIntegrationTest {
     list_nonEmptyBucket_nullPrefix_object(
         (gcs, bucket, prefix) -> getObjectNames(gcs.listObjectInfo(bucket, prefix)),
         (bucket, prefix) ->
-            listRequestWithTrailingDelimiter(
-                bucket, prefix, /* maxResults= */ 1024, /* pageToken= */ null));
+            listRequestWithTrailingDelimiter(bucket, prefix, /* pageToken= */ null));
   }
 
   @Test
@@ -195,8 +190,7 @@ public class GoogleCloudStorageNewIntegrationTest {
             getObjectNames(
                 gcs.listObjectInfoPage(bucket, prefix, /* pageToken= */ null).getItems()),
         (bucket, prefix) ->
-            listRequestWithTrailingDelimiter(
-                bucket, prefix, /* maxResults= */ 1024, /* pageToken= */ null));
+            listRequestWithTrailingDelimiter(bucket, prefix, /* pageToken= */ null));
   }
 
   private void list_nonEmptyBucket_nullPrefix_object(
@@ -222,8 +216,7 @@ public class GoogleCloudStorageNewIntegrationTest {
     list_nonEmptyBucket_nullPrefix_emptyDir(
         (gcs, bucket, prefix) -> getObjectNames(gcs.listObjectInfo(bucket, prefix)),
         (bucket, prefix) ->
-            listRequestWithTrailingDelimiter(
-                bucket, prefix, /* maxResults= */ 1024, /* pageToken= */ null));
+            listRequestWithTrailingDelimiter(bucket, prefix, /* pageToken= */ null));
   }
 
   @Test
@@ -233,8 +226,7 @@ public class GoogleCloudStorageNewIntegrationTest {
             getObjectNames(
                 gcs.listObjectInfoPage(bucket, prefix, /* pageToken= */ null).getItems()),
         (bucket, prefix) ->
-            listRequestWithTrailingDelimiter(
-                bucket, prefix, /* maxResults= */ 1024, /* pageToken= */ null));
+            listRequestWithTrailingDelimiter(bucket, prefix, /* pageToken= */ null));
   }
 
   private void list_nonEmptyBucket_nullPrefix_emptyDir(
@@ -260,8 +252,7 @@ public class GoogleCloudStorageNewIntegrationTest {
     list_nonEmptyBucket_nullPrefix_subDir(
         (gcs, bucket, prefix) -> getObjectNames(gcs.listObjectInfo(bucket, prefix)),
         (bucket, prefix) ->
-            listRequestWithTrailingDelimiter(
-                bucket, prefix, /* maxResults= */ 1024, /* pageToken= */ null));
+            listRequestWithTrailingDelimiter(bucket, prefix, /* pageToken= */ null));
   }
 
   @Test
@@ -271,8 +262,7 @@ public class GoogleCloudStorageNewIntegrationTest {
             getObjectNames(
                 gcs.listObjectInfoPage(bucket, prefix, /* pageToken= */ null).getItems()),
         (bucket, prefix) ->
-            listRequestWithTrailingDelimiter(
-                bucket, prefix, /* maxResults= */ 1024, /* pageToken= */ null));
+            listRequestWithTrailingDelimiter(bucket, prefix, /* pageToken= */ null));
   }
 
   private void list_nonEmptyBucket_nullPrefix_subDir(
@@ -298,8 +288,7 @@ public class GoogleCloudStorageNewIntegrationTest {
     list_nonEmptyBucket_nullPrefix_implicitSubDir(
         (gcs, bucket, prefix) -> getObjectNames(gcs.listObjectInfo(bucket, prefix)),
         (bucket, prefix) ->
-            listRequestWithTrailingDelimiter(
-                bucket, prefix, /* maxResults= */ 1024, /* pageToken= */ null));
+            listRequestWithTrailingDelimiter(bucket, prefix, /* pageToken= */ null));
   }
 
   @Test
@@ -309,8 +298,7 @@ public class GoogleCloudStorageNewIntegrationTest {
             getObjectNames(
                 gcs.listObjectInfoPage(bucket, prefix, /* pageToken= */ null).getItems()),
         (bucket, prefix) ->
-            listRequestWithTrailingDelimiter(
-                bucket, prefix, /* maxResults= */ 1024, /* pageToken= */ null));
+            listRequestWithTrailingDelimiter(bucket, prefix, /* pageToken= */ null));
   }
 
   private void list_nonEmptyBucket_nullPrefix_implicitSubDir(
@@ -336,8 +324,7 @@ public class GoogleCloudStorageNewIntegrationTest {
     list_prefix_doesNotExist(
         (gcs, bucket, prefix) -> getObjectNames(gcs.listObjectInfo(bucket, prefix)),
         (bucket, prefix) ->
-            listRequestWithTrailingDelimiter(
-                bucket, prefix, /* maxResults= */ 1024, /* pageToken= */ null));
+            listRequestWithTrailingDelimiter(bucket, prefix, /* pageToken= */ null));
   }
 
   @Test
@@ -347,8 +334,7 @@ public class GoogleCloudStorageNewIntegrationTest {
             getObjectNames(
                 gcs.listObjectInfoPage(bucket, prefix, /* pageToken= */ null).getItems()),
         (bucket, prefix) ->
-            listRequestWithTrailingDelimiter(
-                bucket, prefix, /* maxResults= */ 1024, /* pageToken= */ null));
+            listRequestWithTrailingDelimiter(bucket, prefix, /* pageToken= */ null));
   }
 
   private void list_prefix_doesNotExist(
@@ -374,8 +360,7 @@ public class GoogleCloudStorageNewIntegrationTest {
     list_prefixObject_empty(
         (gcs, bucket, prefix) -> getObjectNames(gcs.listObjectInfo(bucket, prefix)),
         (bucket, prefix) ->
-            listRequestWithTrailingDelimiter(
-                bucket, prefix, /* maxResults= */ 1024, /* pageToken= */ null));
+            listRequestWithTrailingDelimiter(bucket, prefix, /* pageToken= */ null));
   }
 
   @Test
@@ -385,8 +370,7 @@ public class GoogleCloudStorageNewIntegrationTest {
             getObjectNames(
                 gcs.listObjectInfoPage(bucket, prefix, /* pageToken= */ null).getItems()),
         (bucket, prefix) ->
-            listRequestWithTrailingDelimiter(
-                bucket, prefix, /* maxResults= */ 1024, /* pageToken= */ null));
+            listRequestWithTrailingDelimiter(bucket, prefix, /* pageToken= */ null));
   }
 
   private void list_prefixObject_empty(
@@ -413,8 +397,7 @@ public class GoogleCloudStorageNewIntegrationTest {
     list_prefixObject_withObject(
         (gcs, bucket, prefix) -> getObjectNames(gcs.listObjectInfo(bucket, prefix)),
         (bucket, prefix) ->
-            listRequestWithTrailingDelimiter(
-                bucket, prefix, /* maxResults= */ 1024, /* pageToken= */ null));
+            listRequestWithTrailingDelimiter(bucket, prefix, /* pageToken= */ null));
   }
 
   @Test
@@ -424,8 +407,7 @@ public class GoogleCloudStorageNewIntegrationTest {
             getObjectNames(
                 gcs.listObjectInfoPage(bucket, prefix, /* pageToken= */ null).getItems()),
         (bucket, prefix) ->
-            listRequestWithTrailingDelimiter(
-                bucket, prefix, /* maxResults= */ 1024, /* pageToken= */ null));
+            listRequestWithTrailingDelimiter(bucket, prefix, /* pageToken= */ null));
   }
 
   private void list_prefixObject_withObject(
@@ -451,8 +433,7 @@ public class GoogleCloudStorageNewIntegrationTest {
     list_prefixObject_withSubdir(
         (gcs, bucket, prefix) -> getObjectNames(gcs.listObjectInfo(bucket, prefix)),
         (bucket, prefix) ->
-            listRequestWithTrailingDelimiter(
-                bucket, prefix, /* maxResults= */ 1024, /* pageToken= */ null));
+            listRequestWithTrailingDelimiter(bucket, prefix, /* pageToken= */ null));
   }
 
   @Test
@@ -462,8 +443,7 @@ public class GoogleCloudStorageNewIntegrationTest {
             getObjectNames(
                 gcs.listObjectInfoPage(bucket, prefix, /* pageToken= */ null).getItems()),
         (bucket, prefix) ->
-            listRequestWithTrailingDelimiter(
-                bucket, prefix, /* maxResults= */ 1024, /* pageToken= */ null));
+            listRequestWithTrailingDelimiter(bucket, prefix, /* pageToken= */ null));
   }
 
   private void list_prefixObject_withSubdir(
@@ -489,8 +469,7 @@ public class GoogleCloudStorageNewIntegrationTest {
     list_prefixObject_withImplicitSubdir(
         (gcs, bucket, prefix) -> getObjectNames(gcs.listObjectInfo(bucket, prefix)),
         (bucket, prefix) ->
-            listRequestWithTrailingDelimiter(
-                bucket, prefix, /* maxResults= */ 1024, /* pageToken= */ null));
+            listRequestWithTrailingDelimiter(bucket, prefix, /* pageToken= */ null));
   }
 
   @Test
@@ -500,8 +479,7 @@ public class GoogleCloudStorageNewIntegrationTest {
             getObjectNames(
                 gcs.listObjectInfoPage(bucket, prefix, /* pageToken= */ null).getItems()),
         (bucket, prefix) ->
-            listRequestWithTrailingDelimiter(
-                bucket, prefix, /* maxResults= */ 1024, /* pageToken= */ null));
+            listRequestWithTrailingDelimiter(bucket, prefix, /* pageToken= */ null));
   }
 
   private void list_prefixObject_withImplicitSubdir(
@@ -528,8 +506,7 @@ public class GoogleCloudStorageNewIntegrationTest {
     list_implicitPrefix_withObject(
         (gcs, bucket, prefix) -> getObjectNames(gcs.listObjectInfo(bucket, prefix)),
         (bucket, prefix) ->
-            listRequestWithTrailingDelimiter(
-                bucket, prefix, /* maxResults= */ 1024, /* pageToken= */ null));
+            listRequestWithTrailingDelimiter(bucket, prefix, /* pageToken= */ null));
   }
 
   @Test
@@ -539,8 +516,7 @@ public class GoogleCloudStorageNewIntegrationTest {
             getObjectNames(
                 gcs.listObjectInfoPage(bucket, prefix, /* pageToken= */ null).getItems()),
         (bucket, prefix) ->
-            listRequestWithTrailingDelimiter(
-                bucket, prefix, /* maxResults= */ 1024, /* pageToken= */ null));
+            listRequestWithTrailingDelimiter(bucket, prefix, /* pageToken= */ null));
   }
 
   private void list_implicitPrefix_withObject(
@@ -566,8 +542,7 @@ public class GoogleCloudStorageNewIntegrationTest {
     list_implicitPrefix_withSubdir(
         (gcs, bucket, prefix) -> getObjectNames(gcs.listObjectInfo(bucket, prefix)),
         (bucket, prefix) ->
-            listRequestWithTrailingDelimiter(
-                bucket, prefix, /* maxResults= */ 1024, /* pageToken= */ null));
+            listRequestWithTrailingDelimiter(bucket, prefix, /* pageToken= */ null));
   }
 
   @Test
@@ -577,8 +552,7 @@ public class GoogleCloudStorageNewIntegrationTest {
             getObjectNames(
                 gcs.listObjectInfoPage(bucket, prefix, /* pageToken= */ null).getItems()),
         (bucket, prefix) ->
-            listRequestWithTrailingDelimiter(
-                bucket, prefix, /* maxResults= */ 1024, /* pageToken= */ null));
+            listRequestWithTrailingDelimiter(bucket, prefix, /* pageToken= */ null));
   }
 
   private void list_implicitPrefix_withSubdir(
@@ -604,8 +578,7 @@ public class GoogleCloudStorageNewIntegrationTest {
     list_implicitPrefix_withImplicitSubdir(
         (gcs, bucket, prefix) -> getObjectNames(gcs.listObjectInfo(bucket, prefix)),
         (bucket, prefix) ->
-            listRequestWithTrailingDelimiter(
-                bucket, prefix, /* maxResults= */ 1024, /* pageToken= */ null));
+            listRequestWithTrailingDelimiter(bucket, prefix, /* pageToken= */ null));
   }
 
   @Test
@@ -615,8 +588,7 @@ public class GoogleCloudStorageNewIntegrationTest {
             getObjectNames(
                 gcs.listObjectInfoPage(bucket, prefix, /* pageToken= */ null).getItems()),
         (bucket, prefix) ->
-            listRequestWithTrailingDelimiter(
-                bucket, prefix, /* maxResults= */ 1024, /* pageToken= */ null));
+            listRequestWithTrailingDelimiter(bucket, prefix, /* pageToken= */ null));
   }
 
   private void list_implicitPrefix_withImplicitSubdir(
@@ -722,7 +694,7 @@ public class GoogleCloudStorageNewIntegrationTest {
     assertThat(gcsRequestsTracker.getAllRequestStrings())
         .containsExactly(
             listRequestWithTrailingDelimiter(
-                testBucket, /* prefix= */ null, /* maxResults= */ 1024, /* pageToken= */ null));
+                testBucket, /* prefix= */ null, /* pageToken= */ null));
   }
 
   @Test
@@ -741,7 +713,7 @@ public class GoogleCloudStorageNewIntegrationTest {
     assertThat(gcsRequestsTracker.getAllRequestStrings())
         .containsExactly(
             listRequestWithTrailingDelimiter(
-                testBucket, /* prefix= */ null, /* maxResults= */ 1024, /* pageToken= */ null));
+                testBucket, /* prefix= */ null, /* pageToken= */ null));
   }
 
   @Test
@@ -760,7 +732,7 @@ public class GoogleCloudStorageNewIntegrationTest {
     assertThat(gcsRequestsTracker.getAllRequestStrings())
         .containsExactly(
             listRequestWithTrailingDelimiter(
-                testBucket, /* prefix= */ null, /* maxResults= */ 1024, /* pageToken= */ null));
+                testBucket, /* prefix= */ null, /* pageToken= */ null));
   }
 
   @Test
@@ -778,8 +750,7 @@ public class GoogleCloudStorageNewIntegrationTest {
     assertThat(getObjectNames(listedObjects)).containsExactly(testDir);
     assertThat(gcsRequestsTracker.getAllRequestStrings())
         .containsExactly(
-            listRequestWithTrailingDelimiter(
-                testBucket, testDir, /* maxResults= */ 1024, /* pageToken= */ null));
+            listRequestWithTrailingDelimiter(testBucket, testDir, /* pageToken= */ null));
   }
 
   @Test
@@ -797,8 +768,7 @@ public class GoogleCloudStorageNewIntegrationTest {
     assertThat(getObjectNames(listedObjects)).containsExactly(testDir, testDir + "obj");
     assertThat(gcsRequestsTracker.getAllRequestStrings())
         .containsExactly(
-            listRequestWithTrailingDelimiter(
-                testBucket, testDir, /* maxResults= */ 1024, /* pageToken= */ null));
+            listRequestWithTrailingDelimiter(testBucket, testDir, /* pageToken= */ null));
   }
 
   @Test
@@ -816,8 +786,7 @@ public class GoogleCloudStorageNewIntegrationTest {
     assertThat(getObjectNames(listedObjects)).containsExactly(testDir, testDir + "dir/");
     assertThat(gcsRequestsTracker.getAllRequestStrings())
         .containsExactly(
-            listRequestWithTrailingDelimiter(
-                testBucket, testDir, /* maxResults= */ 1024, /* pageToken= */ null));
+            listRequestWithTrailingDelimiter(testBucket, testDir, /* pageToken= */ null));
   }
 
   @Test
@@ -835,8 +804,7 @@ public class GoogleCloudStorageNewIntegrationTest {
     assertThat(getObjectNames(listedObjects)).containsExactly(testDir, testDir + "obj");
     assertThat(gcsRequestsTracker.getAllRequestStrings())
         .containsExactly(
-            listRequestWithTrailingDelimiter(
-                testBucket, testDir, /* maxResults= */ 1024, /* pageToken= */ null));
+            listRequestWithTrailingDelimiter(testBucket, testDir, /* pageToken= */ null));
   }
 
   @Test
@@ -854,8 +822,7 @@ public class GoogleCloudStorageNewIntegrationTest {
     assertThat(getObjectNames(listedObjects)).containsExactly(testDir, testDir + "dir/");
     assertThat(gcsRequestsTracker.getAllRequestStrings())
         .containsExactly(
-            listRequestWithTrailingDelimiter(
-                testBucket, testDir, /* maxResults= */ 1024, /* pageToken= */ null));
+            listRequestWithTrailingDelimiter(testBucket, testDir, /* pageToken= */ null));
   }
 
   @Test
@@ -889,10 +856,7 @@ public class GoogleCloudStorageNewIntegrationTest {
     assertThat(gcsRequestsTracker.getAllRequestStrings())
         .containsExactly(
             listRequestWithTrailingDelimiter(
-                objectId.getBucketName(),
-                testDirName,
-                /* maxResults= */ 1024,
-                /* pageToken= */ null));
+                objectId.getBucketName(), testDirName, /* pageToken= */ null));
   }
 
   @Test